Move domain management view model test to management package
This also separates out the populated list tests to their own class /
file.
